requirements.txt
pynput
PyQt6
pypinyin
pyyaml
jieba
pyautogui
ollama
